requirements.txt
requests==2.24.0
asn1crypto==1.4.0
pycryptodome==3.9.9
bcrypt==3.2.0
